Simplify package import structure
